# mongodb_dashboard_backend/interfaces/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dashboard API",
    "version": "1.0.0",
    "description": "REST API for Data Management Dashboard with MongoDB and Express"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Include Bearer JWT. Server enforces tenant scoping using tenantId claim."
      }
    },
    "schemas": {
      "GenericDocument": {
        "type": "object",
        "description": "A generic MongoDB document with flexible fields",
        "additionalProperties": true,
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId as string"
          }
        }
      },
      "ListEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenericDocument"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "example": 1
              },
              "limit": {
                "type": "integer",
                "example": 20
              },
              "total": {
                "type": "integer",
                "example": 42
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/app-deployments": {
      "get": {
        "summary": "List application deployments",
        "description": "Paginated list with optional filter and sort.",
        "tags": [
          "AppDeployments"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON filter (e.g., {\"project_id\":\"p1\",\"status\":\"success\"})"
          }
        ],
        "responses": {
          "200": {
            "description": "OK (array or envelope based on pagination params)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GenericDocument"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ListEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter"
          }
        }
      },
      "post": {
        "summary": "Create app deployment",
        "tags": [
          "AppDeployments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/api/llm-costs": {
      "get": {
        "summary": "List LLM cost records",
        "description": "Returns a list of LLM cost documents. Supports optional JSON filter, sorting and pagination. If explicit pagination (page/limit) is provided, response is wrapped with { success, data, meta }. Otherwise a raw array is returned.\n",
        "tags": [
          "LLMCosts"
        ],
        "responses": {
          "200": {
            "description": "Successful response (array or envelope based on pagination params)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GenericDocument"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ListEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Optional page number to enable envelope response"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Optional page size to enable envelope response"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Sort string (e.g., -timestamp or total_cost)"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON filter (e.g., {\"tenant_id\":\"org1\",\"llm_model\":\"gpt-4o\"})"
          }
        ]
      },
      "post": {
        "summary": "Create LLM cost record",
        "tags": [
          "LLMCosts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/api/session/tenants": {
      "get": {
        "summary": "List authorized tenants for current user",
        "description": "Returns the list of tenants from the authenticated user's record (users.tenants). Requires Authorization header.",
        "tags": [
          "Auth",
          "Tenants"
        ],
        "responses": {
          "200": {
            "description": "Authorized tenants for the user"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/session-tracking": {
      "get": {
        "summary": "List session tracking records",
        "description": "Returns a list of session tracking documents. If explicit pagination (page/limit) is provided, the response will be wrapped in an envelope with meta; otherwise, a raw array is returned.\n",
        "tags": [
          "SessionTracking"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Alias for \"limit\" (page size)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON filter (e.g., {\"tenant_id\":\"org1\",\"status\":\"active\"})"
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive text search applied across multiple fields: task_id, tenant_id, organization_name, user_name, project_id, container_id, service_type, status, and session_data fields (session_name, description, llm_model).\n"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response (array or envelope based on pagination params)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GenericDocument"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ListEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter"
          }
        }
      },
      "post": {
        "summary": "Create session tracking record",
        "tags": [
          "SessionTracking"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/api/users/tenant-summary": {
      "get": {
        "summary": "Tenant-wise users summary",
        "description": "Aggregates distinct active users per tenant primarily from the session_tracking collection. Accepts optional time range and status filters. If includeInactive=true, falls back to the users collection joined with known tenants to include tenants without recent activity.\n",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Optional ISO date-time lower bound"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Optional ISO date-time upper bound"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Session status filter. Default \"completed|active\" (i.e., completed or active)."
          },
          {
            "in": "query",
            "name": "includeInactive",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "When true, includes tenants from tenants/users collections even if no activity is found."
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated tenant user counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "tenant_id": {
                            "type": "string"
                          },
                          "tenant_name": {
                            "type": "string",
                            "nullable": true
                          },
                          "user_count": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Number of tenant groups returned"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Database not connected"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "List users",
        "description": "Retrieve a paginated list of users with optional JSON filtering and sorting.",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number (default 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Page size (default 20, max 200)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Sort string (e.g., -created_at)"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string"
            },
            "description": "JSON string filter (e.g., {\"referral_code\":\"ABC\"})"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users (array or envelope based on pagination params)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GenericDocument"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ListEnvelope"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter"
          }
        }
      },
      "post": {
        "summary": "Create user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "User document payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
...
  },
...
}